Add capture_and_plot script; refactor On/Off controller and PID logic for improved On/off controller is functional with hysterisis; PID controller does not function correctly yet

+++ src/filter.hxx
@@ -17,7 +17,7 @@ namespace filter {
         double last_output;       /**< Última muestra suavizada $y(k-1)$ */
     public:
         ExpLowPassFilter(double cutoff_freq  = timing::NYQUIST_FREQ);
-        double operator()(double input);
+        int16_t operator()(double input);
     };
 }
